DeckImporter capable of importing decks from magic.tcgplayer.com
Modified Selector for finding the deck so that it uses more condensed
selectors

Refactored the executions of the MageFileDeckSaver format method to a
utility class.

Added some extra attributes to decks (nothing parses these out yet, was
hoping to do this in TcgPlayerDeckImporter but its format wasn't
friendly and now I'm spent!).

Modified SideBoard and MainBoard so that they extends from CardList.